requirements.txt
torch~=2.3.0
torchvision~=0.18.0
tqdm~=4.66.5
matplotlib~=3.8.0
streamlit~=1.41.1
pillow~=11.0.0